Moved BatchUpload code from separate folder into MercuryServices, fixed changed row bug in Bottles page, made user-requested UI changes to Bromination Samples Search. Finally fixed nav bar's height issue.
