+++ logs/mercury.log
@@ -1,61 +1,1 @@
-[2015-03-03 11:51:34] | WARNING [django.request:143] | Not Found: /browserConnection/buildInfo
-[2015-03-03 11:51:41] | WARNING [django.request:143] | Not Found: /favicon.ico
-[2015-03-03 11:51:44] | WARNING [django.request:143] | Not Found: /startTime
-[2015-03-03 11:53:58] | ERROR [django.request:231] | Internal Server Error: /mercuryservices/fullresults/
-Traceback (most recent call last):
-  File "C:\Python34\lib\site-packages\django\core\handlers\base.py", line 111, in get_response
-    response = wrapped_callback(request, *callback_args, **callback_kwargs)
-  File "C:\Python34\lib\site-packages\django\views\decorators\csrf.py", line 57, in wrapped_view
-    return view_func(*args, **kwargs)
-  File "C:\Python34\lib\site-packages\rest_framework\viewsets.py", line 79, in view
-    return self.dispatch(request, *args, **kwargs)
-  File "C:\Python34\lib\site-packages\rest_framework\views.py", line 403, in dispatch
-    response = self.handle_exception(exc)
-  File "C:\Python34\lib\site-packages\rest_framework\views.py", line 400, in dispatch
-    response = handler(request, *args, **kwargs)
-  File "C:\Python34\lib\site-packages\rest_framework\mixins.py", line 98, in list
-    return Response(serializer.data)
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 576, in data
-    self._data = self.to_native(obj)
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 355, in to_native
-    value = field.field_to_native(obj, field_name)
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 414, in field_to_native
-    return [self.to_native(item) for item in value.all()]
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 414, in <listcomp>
-    return [self.to_native(item) for item in value.all()]
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 350, in to_native
-    for field_name, field in self.fields.items():
-  File "C:\Python34\lib\site-packages\django\utils\functional.py", line 55, in __get__
-    res = instance.__dict__[self.func.__name__] = self.func(instance)
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 217, in fields
-    return self.get_fields()
-  File "C:\Python34\lib\site-packages\rest_framework\serializers.py", line 250, in get_fields
-    new[key] = ret[key]
-KeyError: 'method_code'
-[2015-03-03 11:58:46] | INFO [merlin.views:1347] | POST http://localhost:8000/mercuryauth/login OK 200
-[2015-03-03 11:58:46] | INFO [merlin.views:1364] | GET http://localhost:8000/mercuryservices/users/?username=admin OK 200
-[2015-03-03 11:59:09] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 11:59:09] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 12:01:00] | WARNING [django.request:143] | Not Found: /favicon.ico
-[2015-03-03 12:03:11] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 12:03:11] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 12:03:35] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 12:03:35] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 12:51:59] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 12:51:59] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 12:53:00] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 12:53:00] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 12:58:58] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 12:58:58] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 13:08:32] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 13:08:32] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 13:09:24] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 13:09:24] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 13:09:45] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 13:09:45] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 13:10:57] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 13:10:57] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 13:11:37] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 13:11:37] | INFO [merlin.views:545] | search results count: 1850
-[2015-03-03 13:14:52] | INFO [merlin.views:543] | GET http://localhost:8000/mercuryservices/fullresults/?date_after=2014-08-01&date_before=2014-09-01  OK 200
-[2015-03-03 13:14:52] | INFO [merlin.views:545] | search results count: 1850
+[2015-03-06 11:02:51] | WARNING [django.request:143] | Not Found: /favicon.ico
