+++ mercury/settings.py
@@ -110,7 +110,6 @@ INSTALLED_APPS = (
     'rest_framework_swagger',
     'corsheaders',
     'mercuryservices',
-    #'mercurybatchupload',
     'merlin',
 )
