+++ templates/merlin/acids.html
@@ -7,7 +7,7 @@
             <button type="button" class="btn btn-default" name="add">Add New Acid</button>
         </p>
         <div id="form_add" class="panel-default" style="display:none;">
-            <h3>Create Bottle</h3>
+            <h3>Create Acid</h3>
             <form class="form-inline" style="border:1px solid;width:1000px;padding:4px;">
                 <div class="form-group" style="padding:2px;text-align:center">
                     <label for="code">Code</label><br />
