+++ templates/merlin/base.html
@@ -68,7 +68,7 @@
             <div id="sidebar-wrapper">
                 <ul class="sidebar-nav">
                     <li class="sidebar-brand">
-                        <a href="#">
+                        <a href="javascript:void(0);">
                             Navigation
                         </a>
                     </li>
